remote/requirements.txt
nats-py==2.14.0
psutil==7.2.2
pyalsaaudio==0.11.0
